+++ packages/system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ca-system
 
+## 0.1.0-dev.31
+
+### Patch Changes
+
+- fix: SelectInput paginationMode
+
+- Updated dependencies []:
+  - ca-i18n@0.1.0-dev.31
+
 ## 0.1.0-dev.29
 
 ### Minor Changes